api: Remove runner related infromation from ModelInput.
The goal here is to not leak more information than is absolutely needed
by connectors into the ModelInput and ModelOutput APIs. Previously we
were passing the full input payload through to the connector and other
infromation like model-slug and model-image which were not necessary at
that level. Since we are also validating the schema everywhere this led
to situations where the schema for the model-runner changed in a way
that was irrelevant to the connectors, but they would fail to validate
the new schema even though it was not actually a part of the schema that
was relevant to them.

In order to rectify this I have done the following:
1. Remove the `Model` type from the `ModelInput`.
1. Published a new `runner.json` schema in addition to the `input.json
   and `output.json` schemas.
1. Aligned model-runner and the connectors to validate that correct
   schema.
1. Added an assumption to `model-runner` that it will only ever care
   about the **first** model in the models list. This works because we
   re-write the `inputFile.json` in the matrix part of the workflow in
   the control-plane.