api: Remove ModelSlug enum.
This was no longer being used as an enum now that model connectors have
been made independent and their Docker image is supplied as part of the
payload. However, because it was still being validated as part of the
schema it was making it more difficult to add new models.